Datasources: Collapse the new instance APIs into a single list-item API
Replaces useDataSourceInstanceMeta/useDataSourceInstanceType and their
getters with one getDataSourceInstanceListItem()/useDataSourceInstanceListItem()
pair returning DataSourceInstanceListItem.

The list item already carries uid, type, name, meta, readOnly and isDefault,
so it covers both the type-only and the plugin-metadata call sites with one
API instead of two overlapping ones.

`meta` is still sourced from the plugin meta cache rather than the copy
embedded on the instance settings, with the instance copy kept as a fallback
for runtime-registered datasources that have no entry in that cache.

Part of #125089
