+++ packages/grafana-runtime/src/unstable.ts
@@ -20,26 +20,23 @@ export {
   type GetDataSourceInstanceListFilters,
   getDataSourceInstanceSettings,
   getDataSourceInstanceList,
-  getDataSourceInstanceType,
   getDefaultDataSourceInstance,
   hasDataSourceInstance,
   reloadDataSourceInstanceSettings,
 } from './services/dataSource/settings';
 export { getDataSourceInstance, registerRuntimeDataSourceInstance } from './services/dataSource/dataSource';
-export { getDataSourceInstanceMeta } from './services/dataSource/meta';
+export { getDataSourceInstanceListItem } from './services/dataSource/listItem';
 export {
   useDataSourceInstanceSettings,
   useDataSourceInstance,
   useDataSourceInstanceList,
-  useDataSourceInstanceMeta,
-  useDataSourceInstanceType,
+  useDataSourceInstanceListItem,
   useDefaultDataSourceInstance,
   useHasDataSourceInstance,
   type UseDataSourceInstanceSettingsResult,
   type UseDataSourceInstanceResult,
   type UseDataSourceInstanceListResult,
-  type UseDataSourceInstanceMetaResult,
-  type UseDataSourceInstanceTypeResult,
+  type UseDataSourceInstanceListItemResult,
   type UseDefaultDataSourceInstanceResult,
   type UseHasDataSourceInstanceResult,
 } from './services/dataSource/hooks';
